data-aggregator/pyproject.toml
[tool.poetry]
name = "data-aggregator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
confluent-kafka = "^2.5.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytz = "^2024.1"
aiohttp = "^3.10.5"
beautifulsoup4 = "^4.12.3"
langchain-openai = "^0.2.11"
langchain = "^0.3.9"
langchain-community = "^0.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
